@@ -1,6 +1,6 @@
 using CK.Core;
 using CK.SqlServer;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 
 namespace CK.DB.Actor
 {
