@@ -1,7 +1,7 @@
 using CK.Core;
 using CK.SqlServer;
 using System.Collections.Generic;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Threading;
 using System.Threading.Tasks;
 
