@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using CK.SqlServer;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using CK.Core;
 
 namespace CK.DB.Auth
