@@ -2,7 +2,7 @@ using CK.Core;
 using CK.SqlServer;
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Linq;
 
 namespace CK.DB.Culture
