@@ -1,12 +1,11 @@
 using CK.SqlServer;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using CK.Core;
 using System.Threading;
 using CK.DB.Auth;
-using CK.Text;
 
 namespace CK.DB.User.UserOidc
 {
@@ -119,11 +118,14 @@ namespace CK.DB.User.UserOidc
         /// <param name="sub">The sub that identifies the user in the <paramref name="schemeSuffix"/>.</param>
         /// <param name="cancellationToken">Optional cancellation token.</param>
         /// <returns>A <see cref="IdentifiedUserInfo{T}"/> object or null if not found.</returns>
-        public Task<IdentifiedUserInfo<IUserOidcInfo>> FindKnownUserInfoAsync( ISqlCallContext ctx, string schemeSuffix, string sub, CancellationToken cancellationToken = default( CancellationToken ) )
+        public async Task<IdentifiedUserInfo<IUserOidcInfo?>> FindKnownUserInfoAsync( ISqlCallContext ctx, string schemeSuffix, string sub, CancellationToken cancellationToken = default )
         {
             using( var c = CreateReaderCommand( schemeSuffix, sub ) )
             {
-                return ctx[Database].ExecuteSingleRowAsync( c, r => r == null ? null : DoCreateUserUnfo( schemeSuffix, sub, r ) );
+                return await ctx[Database].ExecuteSingleRowAsync( c, r => r == null 
+                                                                            ? null 
+                                                                            : DoCreateUserUnfo( schemeSuffix, sub, r ), cancellationToken )
+                                           .ConfigureAwait( false );
             }
         }
 
