@@ -1,4 +1,4 @@
-﻿using CK.Core;
+using CK.Core;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,7 +9,7 @@ namespace CK.DB.Auth
     /// <summary>
     /// Helpers methods to convert authentication payload into typed POCO.
     /// </summary>
-    static public class PocoFactotyExtensions
+    static public class PocoFactoryExtensions
     {
         /// <summary>
         /// Extracts payload by first checking whether <paramref name="payload"/> is already a <typeparamref name="T"/>
@@ -23,23 +23,50 @@ namespace CK.DB.Auth
         {
             if( payload == null ) throw new ArgumentNullException( nameof( payload ) );
             if( payload is T ) return (T)payload;
-            var kindOfDic = payload as IEnumerable<KeyValuePair<string, object>>;
-            if( kindOfDic != null ) return ExtractPayload( @this, kindOfDic );
-            throw new ArgumentException( $"Invalid payload. It must be a '{typeof( T ).Name}' POCO or a IEnumerable<KeyValuePair<string, object>>.", nameof( payload ) );
+            var kindOfDic = ToValueTuples( payload );
+            if( kindOfDic == null )
+            {
+                Throw.ArgumentException( nameof( payload ), $"Must be a '{typeof( T ).ToCSharpName()}' POCO, a IEnumerable<(string, object?)> or a IEnumerable<KeyValuePair<string, object?>>." );
+            }
+            return ExtractPayload( @this, kindOfDic );
         }
 
+        internal static IEnumerable<(string Key, object Value)>? ToValueTuples( object payload )
+        {
+            if( payload is not IEnumerable<(string Key, object? Value)> kindOfDic )
+            {
+                if( payload is IEnumerable<KeyValuePair<string, object?>> kv )
+                {
+                    kindOfDic = kv.Select( x => (x.Key, x.Value) );
+                }
+                else if( payload is IEnumerable<KeyValuePair<string, string?>> kvS )
+                {
+                    kindOfDic = kvS.Select( x => (x.Key, (object?)x.Value) );
+                }
+                else if( payload is IEnumerable<(string, string?)> vtS )
+                {
+                    kindOfDic = vtS.Select( x => (x.Item1, (object?)x.Item2) );
+                }
+                else
+                {
+                    kindOfDic = null;
+                }
+            }
+            return kindOfDic;
+        }
+
+
         /// <summary>
-        /// Populates a new instance of <typeparamref name="T"/> with the provided KeyValuePair&lt;string, object&gt;.
+        /// Populates a new instance of <typeparamref name="T"/> with the provided set of (string, object?);.
         /// </summary>
         /// <typeparam name="T">The POCO type.</typeparam>
         /// <param name="this">This POCO factory.</param>
         /// <param name="payload">The payload. Must not be null.</param>
         /// <returns>The resulting POCO.</returns>
-        public static T ExtractPayload<T>(
-            this IPocoFactory<T> @this,
-            IEnumerable<KeyValuePair<string, object>> payload ) where T : IPoco
+        public static T ExtractPayload<T>( this IPocoFactory<T> @this,
+                                           IEnumerable<(string Key, object? Value)> payload ) where T : IPoco
         {
-            if( payload == null ) throw new ArgumentNullException( nameof( payload ) );
+            Throw.CheckNotNullArgument( payload );
             T info = @this.Create();
             var properties = @this.PocoClassType.GetProperties( BindingFlags.Public | BindingFlags.Instance );
             foreach( var kv in payload )
