@@ -128,7 +128,7 @@ namespace CK.DB.Auth.Tests
                         info.Schemes.Count.Should().Be( 1 );
                         info.Schemes[0].Name.Should().StartWith( g.ProviderName );
                         info.Schemes[0].Name.Should().BeEquivalentTo( schemeOrProviderName );
-                        info.Schemes[0].LastUsed.Should().BeCloseTo( DateTime.UtcNow, 1000 );
+                        info.Schemes[0].LastUsed.Should().BeCloseTo( DateTime.UtcNow, TimeSpan.FromSeconds( 1 ) );
 
                         g.DestroyUser( ctx, 1, userId );
                         info = auth.ReadUserAuthInfo( ctx, 1, userId );
@@ -147,7 +147,7 @@ namespace CK.DB.Auth.Tests
                         info.Schemes.Should().HaveCount( 1 );
                         info.Schemes[0].Name.Should().StartWith( g.ProviderName );
                         info.Schemes[0].Name.Should().BeEquivalentTo( schemeOrProviderName );
-                        info.Schemes[0].LastUsed.Should().BeCloseTo( DateTime.UtcNow, 1000 );
+                        info.Schemes[0].LastUsed.Should().BeCloseTo( DateTime.UtcNow, TimeSpan.FromSeconds( 1 ) );
 
                         g.LoginUser( ctx, payloadForLoginFail( userId, userName ) ).UserId.Should().Be( 0 );
 
@@ -183,13 +183,11 @@ namespace CK.DB.Auth.Tests
             }
         }
 
-        static public async Task StandardTestForGenericAuthenticationProviderAsync(
-            Package auth,
-            string schemeOrProviderName,
-            Func<int, string, object> payloadForCreateOrUpdate,
-            Func<int, string, object> payloadForLogin,
-            Func<int, string, object> payloadForLoginFail
-            )
+        static public async Task StandardTestForGenericAuthenticationProviderAsync( Package auth,
+                                                                                    string schemeOrProviderName,
+                                                                                    Func<int, string, object> payloadForCreateOrUpdate,
+                                                                                    Func<int, string, object> payloadForLogin,
+                                                                                    Func<int, string, object> payloadForLoginFail )
         {
             var user = TestHelper.StObjMap.StObjs.Obtain<Actor.UserTable>();
             IGenericAuthenticationProvider g = auth.FindProvider( schemeOrProviderName );
@@ -220,7 +218,7 @@ namespace CK.DB.Auth.Tests
                         info.Schemes.Should().HaveCount( 1 );
                         info.Schemes[0].Name.Should().StartWith( g.ProviderName );
                         info.Schemes[0].Name.Should().BeEquivalentTo( schemeOrProviderName );
-                        info.Schemes[0].LastUsed.Should().BeCloseTo( DateTime.UtcNow, 1000 );
+                        info.Schemes[0].LastUsed.Should().BeCloseTo( DateTime.UtcNow, TimeSpan.FromSeconds( 1 ) );
 
                         await g.DestroyUserAsync( ctx, 1, userId );
                         info = await auth.ReadUserAuthInfoAsync( ctx, 1, userId );
@@ -239,7 +237,7 @@ namespace CK.DB.Auth.Tests
                         info.Schemes.Count.Should().Be( 1 );
                         info.Schemes[0].Name.Should().StartWith( g.ProviderName );
                         info.Schemes[0].Name.Should().BeEquivalentTo( schemeOrProviderName );
-                        info.Schemes[0].LastUsed.Should().BeCloseTo( DateTime.UtcNow, 1000 );
+                        info.Schemes[0].LastUsed.Should().BeCloseTo( DateTime.UtcNow, TimeSpan.FromSeconds( 1 ) );
 
                         (await g.LoginUserAsync( ctx, payloadForLoginFail( userId, userName ) )).UserId.Should().Be( 0 );
 
@@ -267,8 +265,8 @@ namespace CK.DB.Auth.Tests
                     }
                     using( TestHelper.Monitor.OpenInfo( "Invalid payload MUST throw an ArgumentException." ) )
                     {
-                        g.Awaiting( sut => sut.CreateOrUpdateUserAsync( ctx, 1, userId, DBNull.Value ) ).Should().Throw<ArgumentException>();
-                        g.Awaiting( sut => sut.LoginUserAsync( ctx, DBNull.Value ) ).Should().Throw<ArgumentException>();
+                        await g.Awaiting( sut => sut.CreateOrUpdateUserAsync( ctx, 1, userId, DBNull.Value ) ).Should().ThrowAsync<ArgumentException>();
+                        await g.Awaiting( sut => sut.LoginUserAsync( ctx, DBNull.Value ) ).Should().ThrowAsync<ArgumentException>();
                     }
                     using( TestHelper.Monitor.OpenInfo( "Injecting disabled user in sAuthUserOnLogin." ) )
                     using( auth.Database.TemporaryTransform( @"
@@ -304,53 +302,92 @@ namespace CK.DB.Auth.Tests
             var auth = TestHelper.StObjMap.StObjs.Obtain<Package>();
             Assume.That( auth.BasicProvider != null );
 
-            // With Tuple (UserId, Password) payload.  
-            StandardTestForGenericAuthenticationProvider(
-                auth,
-                "Basic",
-                payloadForCreateOrUpdate: ( userId, userName ) => "password",
-                payloadForLogin: ( userId, userName ) => Tuple.Create( userId, "password" ),
-                payloadForLoginFail: ( userId, userName ) => Tuple.Create( userId, "wrong password" ) );
+            // With Tuple<UserId, Password> payload.  
+            StandardTestForGenericAuthenticationProvider( auth,
+                                                         "Basic",
+                                                          payloadForCreateOrUpdate: ( userId, userName ) => "password",
+                                                          payloadForLogin: ( userId, userName ) => Tuple.Create( userId, "password" ),
+                                                          payloadForLoginFail: ( userId, userName ) => Tuple.Create( userId, "wrong password" ) );
 
-            // With Tuple (UserName, Password) payload.  
-            StandardTestForGenericAuthenticationProvider(
-                auth,
-                "Basic",
-                payloadForCreateOrUpdate: ( userId, userName ) => "password",
-                payloadForLogin: ( userId, userName ) => Tuple.Create( userName, "password" ),
-                payloadForLoginFail: ( userId, userName ) => Tuple.Create( userName, "wrong password" ) );
+            // With ValueTuple (UserId, Password) payload.  
+            StandardTestForGenericAuthenticationProvider( auth,
+                                                         "Basic",
+                                                          payloadForCreateOrUpdate: ( userId, userName ) => "password",
+                                                          payloadForLogin: ( userId, userName ) => (userId, "password"),
+                                                          payloadForLoginFail: ( userId, userName ) => (userId, "wrong password") );
 
-            // With KeyValuePairs (UserName, Password) payload.  
-            StandardTestForGenericAuthenticationProvider(
-                auth,
-                "Basic",
-                payloadForCreateOrUpdate: ( userId, userName ) => "£$$µ+",
-                payloadForLogin: ( userId, userName ) => new[]
-                {
-                    new KeyValuePair<string, object>("username", userName),
-                    new KeyValuePair<string, object>("password", "£$$µ+")
-                },
-                payloadForLoginFail: ( userId, userName ) => new[]
-                {
-                    new KeyValuePair<string, object>("username", userName),
-                    new KeyValuePair<string, object>("password", "wrong password")
-                } );
+            // With Tuple<UserName, Password> payload.  
+            StandardTestForGenericAuthenticationProvider( auth,
+                                                          "Basic",
+                                                          payloadForCreateOrUpdate: ( userId, userName ) => "password",
+                                                          payloadForLogin: ( userId, userName ) => Tuple.Create( userName, "password" ),
+                                                          payloadForLoginFail: ( userId, userName ) => Tuple.Create( userName, "wrong password" ) );
 
-            // With KeyValuePairs (UserId, Password) payload.  
-            StandardTestForGenericAuthenticationProvider(
-                auth,
-                "Basic",
-                payloadForCreateOrUpdate: ( userId, userName ) => "MM£$$µ+",
-                payloadForLogin: ( userId, userName ) => new[]
-                {
-                    new KeyValuePair<string, object>("USERID", userId),
-                    new KeyValuePair<string, object>("PASSWORD", "MM£$$µ+")
-                },
-                payloadForLoginFail: ( userId, userName ) => new[]
-                {
-                    new KeyValuePair<string, object>("USERID", userId),
-                    new KeyValuePair<string, object>("PASSWORD", "wrong password")
-                } );
+            // With ValueTuple (UserName, Password) payload.  
+            StandardTestForGenericAuthenticationProvider( auth,
+                                                          "Basic",
+                                                          payloadForCreateOrUpdate: ( userId, userName ) => "password",
+                                                          payloadForLogin: ( userId, userName ) => (userName, "password"),
+                                                          payloadForLoginFail: ( userId, userName ) => (userName, "wrong password") );
+
+            // With KeyValuePairs payload.  
+            StandardTestForGenericAuthenticationProvider( auth,
+                                                          "Basic",
+                                                          payloadForCreateOrUpdate: ( userId, userName ) => "£$$µ+",
+                                                          payloadForLogin: ( userId, userName ) => new[]
+                                                          {
+                                                              new KeyValuePair<string, object>("username", userName),
+                                                              new KeyValuePair<string, object>("password", "£$$µ+")
+                                                          },
+                                                          payloadForLoginFail: ( userId, userName ) => new[]
+                                                          {
+                                                              new KeyValuePair<string, object>("username", userName),
+                                                              new KeyValuePair<string, object>("password", "wrong password")
+                                                          } );
+
+            // With ValueTuples payload.  
+            StandardTestForGenericAuthenticationProvider( auth,
+                                                          "Basic",
+                                                          payloadForCreateOrUpdate: ( userId, userName ) => "£$$µ+",
+                                                          payloadForLogin: ( userId, userName ) => new[]
+                                                          {
+                                                              ("username", userName),
+                                                              ("password", "£$$µ+")
+                                                          },
+                                                          payloadForLoginFail: ( userId, userName ) => new[]
+                                                          {
+                                                              ("username", userName),
+                                                              ("password", "wrong password")
+                                                          } );
+
+            // With KeyValuePairs payload.  
+            StandardTestForGenericAuthenticationProvider( auth,
+                                                          "Basic",
+                                                          payloadForCreateOrUpdate: ( userId, userName ) => "MM£$$µ+",
+                                                          payloadForLogin: ( userId, userName ) => new[]
+                                                          {
+                                                              new KeyValuePair<string, object>("USERID", userId),
+                                                              new KeyValuePair<string, object>("PASSWORD", "MM£$$µ+")
+                                                          },
+                                                          payloadForLoginFail: ( userId, userName ) => new[]
+                                                          {
+                                                              new KeyValuePair<string, object>("USERID", userId),
+                                                              new KeyValuePair<string, object>("PASSWORD", "wrong password")
+                                                          } );
+            // With ValueTuples payload.  
+            StandardTestForGenericAuthenticationProvider( auth,
+                                                          "Basic",
+                                                          payloadForCreateOrUpdate: ( userId, userName ) => "MM£$$µ+",
+                                                          payloadForLogin: ( userId, userName ) => new ValueTuple<string,object?>[]
+                                                          {
+                                                              ("USERID", userId),
+                                                              ("PASSWORD", "MM£$$µ+")
+                                                          },
+                                                          payloadForLoginFail: ( userId, userName ) => new ValueTuple<string, object?>[]
+                                                          {
+                                                              ("USERID", userId),
+                                                              ("PASSWORD", "wrong password")
+                                                          } );
         }
 
         [Test]
@@ -359,21 +396,33 @@ namespace CK.DB.Auth.Tests
             var auth = TestHelper.StObjMap.StObjs.Obtain<Package>();
             Assume.That( auth.BasicProvider != null );
 
-            // With (UserId, Password) payload.  
-            await StandardTestForGenericAuthenticationProviderAsync(
-                auth,
-                "Basic",
-                payloadForCreateOrUpdate: ( userId, userName ) => "password",
-                payloadForLogin: ( userId, userName ) => Tuple.Create( userId, "password" ),
-                payloadForLoginFail: ( userId, userName ) => Tuple.Create( userId, "wrong password" ) );
+            // With Tuple (UserId, Password) payload.  
+            await StandardTestForGenericAuthenticationProviderAsync( auth,
+                                                                     "Basic",
+                                                                     payloadForCreateOrUpdate: ( userId, userName ) => "password",
+                                                                     payloadForLogin: ( userId, userName ) => Tuple.Create( userId, "password" ),
+                                                                     payloadForLoginFail: ( userId, userName ) => Tuple.Create( userId, "wrong password" ) );
 
-            // With (UserName, Password) payload.  
-            await StandardTestForGenericAuthenticationProviderAsync(
-                auth,
-                "Basic",
-                payloadForCreateOrUpdate: ( userId, userName ) => "password",
-                payloadForLogin: ( userId, userName ) => Tuple.Create( userName, "password" ),
-                payloadForLoginFail: ( userId, userName ) => Tuple.Create( userName, "wrong password" ) );
+            // With ValueTuple (UserId, Password) payload.  
+            await StandardTestForGenericAuthenticationProviderAsync( auth,
+                                                                     "Basic",
+                                                                     payloadForCreateOrUpdate: ( userId, userName ) => "password",
+                                                                     payloadForLogin: ( userId, userName ) => (userId, "password"),
+                                                                     payloadForLoginFail: ( userId, userName ) => (userId, "wrong password") );
+
+            // With Tuple (UserName, Password) payload.  
+            await StandardTestForGenericAuthenticationProviderAsync( auth,
+                                                                    "Basic",
+                                                                     payloadForCreateOrUpdate: ( userId, userName ) => "password",
+                                                                     payloadForLogin: ( userId, userName ) => Tuple.Create( userName, "password" ),
+                                                                     payloadForLoginFail: ( userId, userName ) => Tuple.Create( userName, "wrong password" ) );
+
+            // With ValueTuple (UserName, Password) payload.  
+            await StandardTestForGenericAuthenticationProviderAsync( auth,
+                                                                    "Basic",
+                                                                     payloadForCreateOrUpdate: ( userId, userName ) => "password",
+                                                                     payloadForLogin: ( userId, userName ) => (userName, "password"),
+                                                                     payloadForLoginFail: ( userId, userName ) => (userName, "wrong password") );
         }
 
         /// <summary>
